refactor: template field router
