+++ src/trpc/routers/template-field-router/router.ts
@@ -1,0 +1,7 @@
+import { createTRPCRouter } from "@/trpc/api/trpc";
+
+import { addFieldProcedure } from "./procedures/add-fields";
+
+export const templateFieldRouter = createTRPCRouter({
+  add: addFieldProcedure,
+});
